add explicit dtype

+++ tests/tflite_tests/functional/__init__.py
@@ -49,8 +49,8 @@ def make_model():
         flatten = tf.keras.layers.Flatten()(depthwise_conv_output)
 
         # Create the fully connected layer with one unit
-        fc_weights = tf.Variable(tf.random_normal([flatten.shape[1], 1]), name='fc_weights')
-        fc_biases = tf.Variable(tf.random_normal([1]), name='fc_biases')
+        fc_weights = tf.Variable(tf.random_normal([flatten.shape[1], 1], dtype=tf.float32), name='fc_weights')
+        fc_biases = tf.Variable(tf.random_normal([1], dtype=tf.float32), name='fc_biases')
 
         # Initialize the variables
         init_op = tf.global_variables_initializer()
